correction

+++ _binary.c
@@ -12,7 +12,7 @@
 int _b(const char *format, va_list arg, int *pcount)
 {
 	unsigned int num = va_arg(arg, unsigned int);
-	long long unsigned int Tens = 1;
+	long unsigned int Tens = 1;
 	unsigned int result = 0;
 
 	(void)format;
